pyproject.toml
[project]
name = "restapi-test-task-maps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "environs>=14.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
